Move browser JS shim under its own directory
