+++ lib/browser/base64.js
