+++ lib/browser/constants.js
