+++ lib/browser/index.js
