+++ lib/browser/lists.js
