+++ lib/browser/objects.js
