+++ lib/browser/results.js
