+++ lib/browser/rpc.js
