+++ lib/browser/util.js
