+++ lib/index.js
@@ -23,7 +23,7 @@ if (typeof Realm != 'undefined') {
     module.exports = Realm;  // eslint-disable-line no-undef
 } else if (navigator.userAgent) {
     // The userAgent will be defined when running in a browser (such as Chrome debugging mode).
-    module.exports = require('./realm');
+    module.exports = require('./browser');
 } else {
     throw new Error('Missing Realm constructor - please ensure RealmReact framework is included!');
 }
